Move analytics out of util.js
