+++ libs/martech/attributes.js
@@ -1,3 +1,6 @@
+const RE_ALPHANUM = /[^0-9a-z]/gi;
+const RE_TRIM_UNDERSCORE = /^_+|_+$/g;
+
 export function decorateBlockAnalytics(blockEl) {
   blockEl.setAttribute('daa-im', 'true');
   blockEl.setAttribute('daa-lh', [...blockEl.classList].join('|'));
@@ -16,3 +19,17 @@ export function decorateLinkAnalytics(textEl, headings) {
     link.setAttribute('daa-ll', str);
   });
 }
+
+export const analyticsGetLabel = (txt) => txt.replaceAll('&', 'and')
+  .replace(RE_ALPHANUM, '_')
+  .replace(RE_TRIM_UNDERSCORE, '');
+
+export const analyticsDecorateList = (li, idx) => {
+  const link = li.firstChild?.nodeName === 'A' && li.firstChild;
+  if (!link) return;
+
+  const label = link.textContent || link.getAttribute('aria-label');
+  if (!label) return;
+
+  link.setAttribute('daa-ll', `${analyticsGetLabel(label)}-${idx + 1}`);
+};
